+++ fe/src/Components/Badge/CategoryBadge.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Colors from '@/Constants/Colors';
+import { Colors } from '@/Constants';
 
 const Badge = styled.button`
   padding: 8px 16px;
