+++ fe/src/Components/Badge/DiscountBadge.jsx
@@ -1,7 +1,6 @@
 import styled from 'styled-components';
 
-import Colors from '@/Constants/Colors';
-import Fonts from '@/Constants/Fonts';
+import { Colors, Fonts } from '@/Constants';
 
 const getBadgeName = type => {
   switch (type) {
