+++ fe/src/Components/BestProducts/BestProducts.jsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 
+import { Colors, Fonts } from '@/Constants';
+
 import { TabList } from './TabList';
 
 import { CategoryBadge } from '@/Components/Badge/CategoryBadge';
 import { ProductCard } from '@/Components/ProductCard';
-import Colors from '@/Constants/Colors';
-import Fonts from '@/Constants/Fonts';
 import { fetchData } from '@/Utils/Utils';
 
 const BestProductWrapper = styled.div`
