+++ fe/src/Components/Button/RetryButton.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Colors from '@/Constants/Colors';
+import { Colors } from '@/Constants';
 
 const Button = styled.button`
   display: flex;
