+++ fe/src/Components/CategoryProductsList/CategoryProducts.jsx
@@ -64,3 +64,4 @@ export const CategoryProducts = props => {
     </CategoryProductsWrapper>
   );
 };
+
