+++ fe/src/Components/LoadingSpinner/index.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Colors from '@/Constants/Colors';
+import { Colors } from '@/Constants';
 
 const LoadingSpinnerWrapper = styled.div`
   width: 120px;
