+++ fe/src/Constants/index.js
@@ -1,1 +1,3 @@
 export { IconFonts } from './IconFonts';
+export { Colors } from './Colors';
+export { Fonts } from './Fonts';
